Updated comments in .properties.
The comments for the property 'first.and.third.day' incorrectly
indicated that this property was used to determine what day to send the
emails.  This is not so.  Instead it is used to identify the day of the
week the meeting occurs on, and is used for calculating the 'first and
third'-ness of that date.